Add arbitrary object to Test event metadata
