+++ platforms/software/testing-base/src/main/java/org/gradle/api/internal/tasks/testing/DefaultTestEventReporter.java
@@ -16,7 +16,6 @@
 
 package org.gradle.api.internal.tasks.testing;
 
-import com.google.common.base.Preconditions;
 import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.tasks.testing.results.DefaultTestResult;
 import org.gradle.api.internal.tasks.testing.results.TestListenerInternal;
@@ -25,15 +24,9 @@ import org.gradle.api.tasks.testing.TestFailure;
 import org.gradle.api.tasks.testing.TestFailureDetails;
 import org.gradle.api.tasks.testing.TestOutputEvent;
 import org.gradle.api.tasks.testing.TestResult;
-import org.gradle.model.internal.manage.schema.extract.ScalarTypes;
-import org.gradle.model.internal.type.ModelType;
 
-import javax.annotation.Nullable;
-import java.io.Serializable;
 import java.time.Instant;
-import java.util.Collection;
 import java.util.Collections;
-import java.util.Map;
 
 @NonNullApi
 class DefaultTestEventReporter implements TestEventReporter {
@@ -73,42 +66,10 @@ class DefaultTestEventReporter implements TestEventReporter {
     }
 
     @Override
-    public void metadata(Instant logTime, Map<String, Object> metadata) {
-        Preconditions.checkNotNull(logTime, "logTime can not be null!");
-        Preconditions.checkNotNull(metadata, "Metadata can not be null!");
-        checkAllowableTypes(metadata);
-
+    public void metadata(Instant logTime, Object metadata) {
         listener.metadata(testDescriptor, new DefaultTestMetadataEvent(logTime.toEpochMilli(), metadata));
     }
 
-    private static void checkAllowableTypes(Map<String, Object> metadata) {
-        if (metadata.entrySet().stream().anyMatch(entry -> entry.getKey() == null)) {
-            throw new IllegalArgumentException("Metadata key cannot be null");
-        }
-        metadata.entrySet().stream().filter(entry -> !isAllowedType(entry.getValue())).findFirst().ifPresent(entry -> {
-            if (entry.getValue() == null) {
-                throw new IllegalArgumentException(String.format("Metadata '%s' has null value", entry.getKey()));
-            } else {
-                throw new IllegalArgumentException(String.format("Metadata '%s' has unsupported value type '%s'", entry.getKey(), entry.getValue().getClass().getName()));
-            }
-        });
-    }
-
-    private static boolean isAllowedType(@Nullable Object value) {
-        if (value == null) {
-            return false;
-        }
-
-        if (ScalarTypes.isScalarType(ModelType.of(value.getClass()))) {
-            return true;
-        }
-
-        if (value instanceof Collection && value instanceof Serializable) {
-            return ((Collection<?>) value).stream().allMatch(DefaultTestEventReporter::isAllowedType);
-        }
-
-        return false;
-    }
 
     @Override
     public void succeeded(Instant endTime) {
