+++ platforms/software/testing-base/src/main/java/org/gradle/api/internal/tasks/testing/LifecycleTrackingTestEventReporter.java
@@ -21,7 +21,6 @@ import org.gradle.api.tasks.testing.TestEventReporter;
 import org.gradle.api.tasks.testing.TestOutputEvent;
 
 import java.time.Instant;
-import java.util.Map;
 
 @NonNullApi
 class LifecycleTrackingTestEventReporter<T extends TestEventReporter> implements TestEventReporter {
@@ -53,7 +52,7 @@ class LifecycleTrackingTestEventReporter<T extends TestEventReporter> implements
     }
 
     @Override
-    public void metadata(Instant logTime, Map<String, Object> metadata) {
+    public void metadata(Instant logTime, Object metadata) {
         requireRunning();
         delegate.metadata(logTime, metadata);
     }
